feat: implement schema for library, info & topic pages
