+++ templates/blog/blog.js
@@ -1,43 +1,5 @@
-import { buildBlock, getMetadata } from '../../scripts/lib-franklin.js';
-import { makePublicUrl, setJsonLd } from '../../scripts/scripts.js';
-
-function buildJsonLd() {
-  const data = {
-    '@context': 'http://schema.org',
-    '@type': 'Article',
-    '@id': `https://lifesciences.danaher.com${makePublicUrl(window.location.pathname)}`,
-    headline: getMetadata('og:title'),
-    image: getMetadata('og:image'),
-    datePublished: getMetadata('publishdate'),
-    publisher: {
-      '@type': 'Organization',
-      name: 'Danaher Life Sciences',
-      logo: {
-        '@type': 'ImageObject',
-        url: 'https://lifesciences.danaher.com/content/dam/danaher/brand-logos/danaher/Logo.svg',
-      },
-    },
-    description: getMetadata('description'),
-    mainEntityOfPage: {
-      '@type': 'WebPage',
-      '@id': `https://lifesciences.danaher.com${makePublicUrl(window.location.pathname)}`,
-    },
-  };
-
-  if (getMetadata('creationdate')) data.datePublished = getMetadata('creationdate');
-  if (getMetadata('updatedate')) data.dateModified = getMetadata('updatedate');
-  if (getMetadata('authorname')) {
-    data.author = {
-      '@type': 'Person',
-      name: getMetadata('authorname'),
-    };
-  }
-
-  setJsonLd(
-    data,
-    'article',
-  );
-}
+import { buildBlock } from '../../scripts/lib-franklin.js';
+import { buildArticleSchema } from '../../scripts/schema.js';
 
 export default async function buildAutoBlocks() {
   const main = document.querySelector('main');
@@ -88,5 +50,5 @@ export default async function buildAutoBlocks() {
     buildBlock('related-articles', { elems: [] }),
   );
 
-  buildJsonLd();
+  buildArticleSchema();
 }
