+++ templates/library/library.js
@@ -1,0 +1,5 @@
+import { buildArticleSchema } from '../../scripts/schema.js';
+
+export default async function buildAutoBlocks() {
+  buildArticleSchema();
+}
