+++ templates/topic/topic.js
@@ -1,5 +1,6 @@
 import { div } from '../../scripts/dom-builder.js';
 import { buildBlock } from '../../scripts/lib-franklin.js';
+import { buildArticleSchema } from '../../scripts/schema.js';
 
 export default async function buildAutoBlocks() {
   const main = document.querySelector('main');
@@ -8,4 +9,6 @@ export default async function buildAutoBlocks() {
   main.firstElementChild.insertAdjacentElement('afterend', sideNavBlock);
   main.querySelector(':scope > div:nth-child(3)')?.prepend(buildBlock('social-media', { elems: [] }));
   main.lastElementChild.append(buildBlock('social-media', { elems: [] }));
+
+  buildArticleSchema();
 }
